Give Settings sections, and move the measurement knobs off the front page
Settings was one unlabelled run of twenty-odd rows. The cards existed but
had no titles, so the first row of each card was doing duty as its
heading — a "Gestures" row that looks exactly like a tappable setting,
sitting directly above four sliders; the same for "Experimental" and
"Routing engine". Nothing on screen distinguished a label from a control,
which is most of why it read as a wall of switches.

Two changes, and neither removes a control.

Sections have names now. `SettingsSection` puts a real heading above each
card: Appearance, Map, General, Offline maps. `SettingsGroup` stays for
the untitled card at the bottom, where the two navigation rows live and a
heading would name a category that does not exist.

And the controls that exist for measuring the app rather than using it —
gesture tunables in milliseconds and degrees, the unfinished map layers,
the routing-engine override with its pack-source URL field and solve
readout — moved behind an Advanced screen. They are still in the shipped
build and still one tap away, deliberately: whether the phone can route
is a property of the RELEASE APK on real silicon, and a debug build
proves none of it. They are just no longer the thing you scroll past
looking for Units.

Build-packs-on-device did not move. It is not a measurement knob — it
changes what tapping Download does, by minutes — so it sits in Offline
maps next to Wi-Fi-only, which is the other setting that changes what a
download means.

The shared composables moved to SettingsCommon.kt so both screens can use
them, and `formatSize` went with the dead code it was: nothing called it.

Tests pin both halves of the move — that everything arrived on Advanced,
and that it is genuinely absent from the main screen rather than
duplicated, which is the half that is easy to skip.
